# src/app/api/verticals/route.ts
import { createClient } from '@/lib/supabase/server'
import { NextRequest, NextResponse } from 'next/server'
import { ApiError, ERROR_CODES } from '@/lib/utils/errors'

export async function GET(request: NextRequest) {
  try {
    const supabase = await createClient()
    const { searchParams } = new URL(request.url)
    
    const organizationId = searchParams.get('organization_id')
    const sector = searchParams.get('sector')
    const tier = searchParams.get('tier')
    const limit = parseInt(searchParams.get('limit') || '100')
    const offset = parseInt(searchParams.get('offset') || '0')

    if (!organizationId) {
      throw new ApiError(
        'Organization ID is required',
        ERROR_CODES.VALIDATION_ERROR,
        400
      )
    }

    let query = supabase
      .from('verticals')
      .select('*')
      .eq('organization_id', organizationId)
      .is('deleted_at', null)

    // Apply filters
    if (sector) query = query.eq('sector', sector)
    if (tier) query = query.eq('tier', tier)

    // Apply pagination
    query = query.range(offset, offset + limit - 1)

    const { data, error, count } = await query

    if (error) {
      throw new ApiError(
        'Failed to fetch verticals',
        ERROR_CODES.DATABASE_ERROR,
        500,
        error
      )
    }

    return NextResponse.json({
      data,
      success: true,
      count
    })

  } catch (error) {
    if (error instanceof ApiError) {
      return NextResponse.json(
        {
          error: error.message,
          code: error.code,
          success: false,
          details: error.details
        },
        { status: error.statusCode }
      )
    }

    return NextResponse.json(
      {
        error: 'Internal server error',
        code: ERROR_CODES.SERVER_ERROR,
        success: false
      },
      { status: 500 }
    )
  }
}

export async function POST(request: NextRequest) {
  try {
    const supabase = await createClient()
    const body = await request.json()

    const {
      organization_id,
      name,
      sector,
      b2b_b2c,
      naics_code,
      revenue_floor,
      typical_revenue_range,
      typical_marketing_budget_pct,
      key_decision_maker_title,
      tier,
      notes
    } = body

    // Validation
    if (!organization_id || !name) {
      throw new ApiError(
        'Organization ID and name are required',
        ERROR_CODES.VALIDATION_ERROR,
        400
      )
    }

    // Check for duplicate names within organization
    const { data: existing, error: checkError } = await supabase
      .from('verticals')
      .select('id')
      .eq('organization_id', organization_id)
      .eq('name', name)
      .is('deleted_at', null)
      .single()

    if (checkError && checkError.code !== 'PGRST116') {
      throw new ApiError(
        'Failed to check for duplicate vertical',
        ERROR_CODES.DATABASE_ERROR,
        500,
        checkError
      )
    }

    if (existing) {
      throw new ApiError(
        'Vertical with this name already exists',
        ERROR_CODES.DUPLICATE_ENTRY,
        409,
        { field: 'name', value: name }
      )
    }

    // Insert new vertical
    const { data, error } = await supabase
      .from('verticals')
      .insert({
        organization_id,
        name,
        sector,
        b2b_b2c,
        naics_code,
        revenue_floor,
        typical_revenue_range,
        typical_marketing_budget_pct,
        key_decision_maker_title,
        tier,
        notes
      })
      .select()
      .single()

    if (error) {
      throw new ApiError(
        'Failed to create vertical',
        ERROR_CODES.DATABASE_ERROR,
        500,
        error
      )
    }

    return NextResponse.json({
      data,
      success: true
    }, { status: 201 })

  } catch (error) {
    if (error instanceof ApiError) {
      return NextResponse.json(
        {
          error: error.message,
          code: error.code,
          success: false,
          details: error.details
        },
        { status: error.statusCode }
      )
    }

    return NextResponse.json(
      {
        error: 'Internal server error',
        code: ERROR_CODES.SERVER_ERROR,
        success: false
      },
      { status: 500 }
    )
  }
}