# src/lib/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OrbitABM API",
    "description": "RESTful API for OrbitABM - Account-Based Marketing Campaign Intelligence Platform",
    "version": "1.0.0",
    "contact": {
      "name": "OrbitABM Support",
      "email": "support@orbitabm.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://orbitabm.com/license"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    },
    {
      "url": "https://your-domain.com/api",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/verticals": {
      "get": {
        "tags": ["Verticals"],
        "summary": "List verticals",
        "description": "Retrieve a list of industry verticals with optional filtering",
        "parameters": [
          {
            "name": "organization_id",
            "in": "query",
            "required": true,
            "description": "Organization UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "sector",
            "in": "query",
            "description": "Filter by sector",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tier",
            "in": "query",
            "description": "Filter by tier",
            "schema": {
              "type": "string",
              "enum": ["tier_1", "tier_2", "tier_3", "borderline", "eliminated"]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of results per page",
            "schema": {
              "type": "integer",
              "default": 100,
              "minimum": 1,
              "maximum": 1000
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of results to skip",
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Vertical"
                      }
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "count": {
                      "type": "integer",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": ["Verticals"],
        "summary": "Create vertical",
        "description": "Create a new industry vertical",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerticalCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Vertical created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Vertical"
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/generated-documents": {
      "get": {
        "tags": ["Generated Documents"],
        "summary": "List generated documents",
        "description": "Retrieve a list of generated documents with optional filtering and pagination",
        "parameters": [
          {
            "$ref": "#/components/parameters/OrganizationId"
          },
          {
            "name": "document_type",
            "in": "query",
            "description": "Filter by document type",
            "schema": {
              "type": "string",
              "enum": ["prospect_research", "campaign_sequence", "competitive_analysis", "audit_report", "proposal"]
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by document status",
            "schema": {
              "type": "string",
              "enum": ["draft", "in_review", "approved", "delivered", "archived"]
            }
          },
          {
            "name": "company_id",
            "in": "query",
            "description": "Filter by company UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "campaign_id",
            "in": "query",
            "description": "Filter by campaign UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/GeneratedDocument"
                      }
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "count": {
                      "type": "integer",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": ["Generated Documents"],
        "summary": "Create generated document",
        "description": "Create a new generated document",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneratedDocumentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Generated document created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/GeneratedDocument"
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/generated-documents/{id}": {
      "get": {
        "tags": ["Generated Documents"],
        "summary": "Get generated document by ID",
        "description": "Retrieve a single generated document by its UUID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Generated document UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/GeneratedDocument"
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "patch": {
        "tags": ["Generated Documents"],
        "summary": "Update generated document",
        "description": "Update an existing generated document (partial update)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Generated document UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneratedDocumentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generated document updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/GeneratedDocument"
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": ["Generated Documents"],
        "summary": "Delete generated document",
        "description": "Soft delete a generated document (sets deleted_at timestamp)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Generated document UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Generated document deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "deleted": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Vertical": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Vertical UUID"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid",
            "description": "Organization UUID"
          },
          "name": {
            "type": "string",
            "description": "Vertical name",
            "example": "HVAC Companies"
          },
          "sector": {
            "type": "string",
            "nullable": true,
            "description": "Industry sector",
            "example": "Home Services"
          },
          "b2b_b2c": {
            "type": "string",
            "enum": ["B2B", "B2C", "Both"],
            "nullable": true,
            "description": "Business model",
            "example": "B2C"
          },
          "naics_code": {
            "type": "string",
            "nullable": true,
            "description": "NAICS industry code",
            "example": "238220"
          },
          "revenue_floor": {
            "type": "number",
            "nullable": true,
            "description": "Minimum revenue threshold",
            "example": 2000000
          },
          "typical_revenue_range": {
            "type": "string",
            "nullable": true,
            "description": "Typical revenue range",
            "example": "$2M - $10M"
          },
          "typical_marketing_budget_pct": {
            "type": "string",
            "nullable": true,
            "description": "Typical marketing budget percentage",
            "example": "3-5%"
          },
          "key_decision_maker_title": {
            "type": "string",
            "nullable": true,
            "description": "Key decision maker title",
            "example": "Owner/CEO"
          },
          "tier": {
            "type": "string",
            "enum": ["tier_1", "tier_2", "tier_3", "borderline", "eliminated"],
            "nullable": true,
            "description": "Vertical tier",
            "example": "tier_1"
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "description": "Additional notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Deletion timestamp"
          }
        },
        "required": ["id", "organization_id", "name"]
      },
      "VerticalCreate": {
        "type": "object",
        "properties": {
          "organization_id": {
            "type": "string",
            "format": "uuid",
            "description": "Organization UUID"
          },
          "name": {
            "type": "string",
            "description": "Vertical name",
            "example": "HVAC Companies"
          },
          "sector": {
            "type": "string",
            "description": "Industry sector",
            "example": "Home Services"
          },
          "b2b_b2c": {
            "type": "string",
            "enum": ["B2B", "B2C", "Both"],
            "description": "Business model"
          },
          "naics_code": {
            "type": "string",
            "description": "NAICS industry code"
          },
          "revenue_floor": {
            "type": "number",
            "description": "Minimum revenue threshold"
          },
          "typical_revenue_range": {
            "type": "string",
            "description": "Typical revenue range"
          },
          "typical_marketing_budget_pct": {
            "type": "string",
            "description": "Typical marketing budget percentage"
          },
          "key_decision_maker_title": {
            "type": "string",
            "description": "Key decision maker title"
          },
          "tier": {
            "type": "string",
            "enum": ["tier_1", "tier_2", "tier_3", "borderline", "eliminated"],
            "description": "Vertical tier"
          },
          "notes": {
            "type": "string",
            "description": "Additional notes"
          }
        },
        "required": ["organization_id", "name"]
      },
...
      "GeneratedDocument": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "description": "Generated document UUID" },
          "organization_id": { "type": "string", "format": "uuid", "description": "Organization UUID" },
          "document_template_id": { "type": "string", "format": "uuid", "nullable": true, "description": "Document template UUID" },
          "company_id": { "type": "string", "format": "uuid", "nullable": true, "description": "Company UUID" },
          "campaign_id": { "type": "string", "format": "uuid", "nullable": true, "description": "Campaign UUID" },
          "title": { "type": "string", "description": "Document title", "example": "Acme HVAC - Prospect Research" },
          "document_type": { "type": "string", "enum": ["prospect_research", "campaign_sequence", "competitive_analysis", "audit_report", "proposal"], "description": "Document type" },
          "status": { "type": "string", "enum": ["draft", "in_review", "approved", "delivered", "archived"], "description": "Document status", "example": "draft" },
          "content": { "type": "object", "description": "Document content (JSON)" },
          "readiness_score": { "type": "integer", "minimum": 1, "maximum": 10, "nullable": true, "description": "Readiness score (1-10)" },
          "version": { "type": "integer", "description": "Document version", "example": 1 },
          "approved_by": { "type": "string", "format": "uuid", "nullable": true, "description": "Approver profile UUID" },
          "approved_at": { "type": "string", "format": "date-time", "nullable": true, "description": "Approval timestamp" },
          "last_generated_at": { "type": "string", "format": "date-time", "nullable": true, "description": "Last generation timestamp" },
          "created_at": { "type": "string", "format": "date-time", "description": "Creation timestamp" },
          "updated_at": { "type": "string", "format": "date-time", "description": "Last update timestamp" },
          "deleted_at": { "type": "string", "format": "date-time", "nullable": true, "description": "Deletion timestamp" }
        },
        "required": ["id", "organization_id", "title", "document_type", "status"]
      },
      "GeneratedDocumentCreate": {
        "type": "object",
        "properties": {
          "organization_id": { "type": "string", "format": "uuid", "description": "Organization UUID" },
          "document_template_id": { "type": "string", "format": "uuid", "description": "Document template UUID" },
          "company_id": { "type": "string", "format": "uuid", "description": "Company UUID" },
          "campaign_id": { "type": "string", "format": "uuid", "description": "Campaign UUID" },
          "title": { "type": "string", "description": "Document title", "example": "Acme HVAC - Prospect Research" },
          "document_type": { "type": "string", "enum": ["prospect_research", "campaign_sequence", "competitive_analysis", "audit_report", "proposal"], "description": "Document type" },
          "status": { "type": "string", "enum": ["draft", "in_review", "approved", "delivered", "archived"], "default": "draft", "description": "Document status" },
          "content": { "type": "object", "description": "Document content (JSON)" },
          "readiness_score": { "type": "integer", "minimum": 1, "maximum": 10, "description": "Readiness score (1-10)" },
          "version": { "type": "integer", "default": 1, "description": "Document version" }
        },
        "required": ["organization_id", "title", "document_type"]
      },
      "GeneratedDocumentUpdate": {
        "type": "object",
        "properties": {
          "document_template_id": { "type": "string", "format": "uuid", "nullable": true, "description": "Document template UUID" },
          "company_id": { "type": "string", "format": "uuid", "nullable": true, "description": "Company UUID" },
          "campaign_id": { "type": "string", "format": "uuid", "nullable": true, "description": "Campaign UUID" },
          "title": { "type": "string", "description": "Document title" },
          "document_type": { "type": "string", "enum": ["prospect_research", "campaign_sequence", "competitive_analysis", "audit_report", "proposal"], "description": "Document type" },
          "status": { "type": "string", "enum": ["draft", "in_review", "approved", "delivered", "archived"], "description": "Document status" },
          "content": { "type": "object", "description": "Document content (JSON)" },
          "readiness_score": { "type": "integer", "minimum": 1, "maximum": 10, "nullable": true, "description": "Readiness score (1-10)" },
          "version": { "type": "integer", "description": "Document version" },
          "approved_by": { "type": "string", "format": "uuid", "nullable": true, "description": "Approver profile UUID" },
          "approved_at": { "type": "string", "format": "date-time", "nullable": true, "description": "Approval timestamp" },
          "last_generated_at": { "type": "string", "format": "date-time", "nullable": true, "description": "Last generation timestamp" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "Company with this name already exists"
          },
          "code": {
            "type": "string",
            "description": "Error code",
            "enum": ["VALIDATION_ERROR", "NOT_FOUND", "DUPLICATE_ENTRY", "FORBIDDEN", "DATABASE_ERROR", "SERVER_ERROR", "NETWORK_ERROR"],
            "example": "DUPLICATE_ENTRY"
          },
          "success": {
            "type": "boolean",
            "example": false
          },
          "details": {
            "type": "object",
            "nullable": true,
            "description": "Additional error details",
            "example": {
              "field": "name",
              "value": "Acme HVAC"
            }
          }
        },
        "required": ["error", "code", "success"]
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad Request - Invalid input data",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Organization ID is required",
              "code": "VALIDATION_ERROR",
              "success": false
            }
          }
        }
      },
      "NotFound": {
        "description": "Not Found - Resource does not exist",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Company not found",
              "code": "NOT_FOUND",
              "success": false
            }
          }
        }
      },
      "Conflict": {
        "description": "Conflict - Resource already exists or constraint violation",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Company with this name already exists",
              "code": "DUPLICATE_ENTRY",
              "success": false,
              "details": {
                "field": "name",
                "value": "Acme HVAC"
              }
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Internal Server Error - Unexpected server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Internal server error",
              "code": "SERVER_ERROR",
              "success": false
            }
          }
        }
      }
    },
    "parameters": {
      "OrganizationId": {
        "name": "organization_id",
        "in": "query",
        "required": true,
        "description": "Organization UUID",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "description": "Number of results per page",
        "schema": {
          "type": "integer",
          "default": 100,
          "minimum": 1,
          "maximum": 1000
        }
      },
      "Offset": {
        "name": "offset",
        "in": "query",
        "description": "Number of results to skip",
        "schema": {
          "type": "integer",
          "default": 0,
          "minimum": 0
        }
      }
    }
  }
}